refactor: simplify `Callable` and subclasses
Let the subclases handle arity checks, since many functions accept
variable number of auguments.

+++ simpl/error.cc
@@ -10,14 +10,16 @@ namespace simpl {
 static bool gHadError = false;
 static bool gHadRuntimeError = false;
 
-void Error(int line, const std::string &message) { Report(line, "", message); }
+void Error(size_t line, const std::string &message) {
+  Report(line, "", message);
+}
 
 void HandleRuntimeError(const std::runtime_error &error) {
   std::cout << error.what() << std::endl;
   gHadRuntimeError = true;
 }
 
-void Report(int line, const std::string &where, const std::string &message) {
+void Report(size_t line, const std::string &where, const std::string &message) {
   std::cout << "[line " << line << "] Error " << where << ": " << message
             << std::endl;
   gHadError = true;
