+++ simpl/error.h
@@ -8,9 +8,9 @@
 
 namespace simpl {
 
-void Error(int line, const std::string &message);
+void Error(size_t line, const std::string &message);
 
-void Report(int line, const std::string &where, const std::string &message);
+void Report(size_t line, const std::string &where, const std::string &message);
 
 void HandleRuntimeError(const std::runtime_error &error);
